[Evals] Add support for OMNI Math (#97)
Adds OMNI-Math. Part 1 of resolving #51 . 


FOr OMNI-Math, the authors did LLM as a Judge evaluation with GPT-4o, but also provided a benchmark subset for rule-based evlauation. We do not support LLM as a Judge yet, so this PR only adds support for rule-based evaluation. 


The dataset for rule-based evaluation is provided in this repo: https://github.com/KbsdJames/omni-math-rule 

I have uploaded the same as an alternate split in the official HF dataset: https://huggingface.co/datasets/KbsdJames/Omni-MATH/discussions/2